+++ modules/@ergonode/attributes/src/components/Forms/AttributeForm.vue
@@ -42,7 +42,6 @@
                     :multiselect="true"
                     :searchable="true"
                     :clearable="true"
-                    :sticky-search="true"
                     :disabled="!isAllowedToUpdate"
                     :error-messages="errors[groupsFieldKey]"
                     href="attributes/groups/autocomplete"