+++ modules/@ergonode/categories/src/extends/store/product/state.js
@@ -2,4 +2,6 @@
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE for license details.
  */
-export default () => ({});
+export default () => ({
+    categories: [],
+});