+++ modules/@ergonode/media/src/components/Modal/ModalMediaTabBar.vue
@@ -5,7 +5,7 @@
 <template>
     <ModalTabBar
         class="modal-media-tab-bar"
-        title="Add images"
+        title="Add resources"
         :items="tabs"
         @close="onCloseModal" />
 </template>