+++ modules/@ergonode/ui/src/components/Input/InputSolidStyle.vue
@@ -146,7 +146,8 @@ export default {
             border: $BORDER_1_GREY;
             box-sizing: border-box;
             background-color: $WHITE;
-            transition: 0.3s cubic-bezier(0.25, 0.8, 0.5, 1);
+            transition: border-color 0.3s cubic-bezier(0.25, 0.8, 0.5, 1);
+            will-change: border-color;
 
             &::after {
                 position: absolute;